Move get_source_addr_v4
